Invites

+++ src/Pages/Advisor/Invites/Invites.jsx
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from 'react'
 import emailjs from '@emailjs/browser'
 import './Invites.css'
 
-const defaultCoinbaseURL =
+const coinbaseURL =
 	'https://www.coinbase.com/oauth/authorize?account=all&scope=wallet:accounts:read,wallet:accounts:update,wallet:accounts:create,wallet:accounts:delete,wallet:addresses:read,wallet:addresses:create,wallet:buys:read,wallet:buys:create,wallet:deposits:read,wallet:deposits:create,wallet:notifications:read,wallet:payment-methods:read,wallet:payment-methods:delete,wallet:payment-methods:limits,wallet:sells:read,wallet:sells:create,wallet:transactions:read,wallet:transactions:request,wallet:transactions:transfer,wallet:user:read,wallet:user:update,wallet:user:email,wallet:withdrawals:read,wallet:withdrawals:create&response_type=code&client_id=d41fd296f21919731b07190afcf278b4f7a7f2813d029bb48d81fbf872c8fae4'
 
 function defaultMessage(advisor) {
@@ -30,7 +30,6 @@ function defaultSubject(advisor) {
 }
 
 const Invites = ({ advisor }) => {
-	const [coinbaseURL] = useState(`${defaultCoinbaseURL}&state=${advisor.idToken.payload.sub}`)
 	const [email, setEmail] = useState('')
 	const [isLoading, setIsLoading] = useState(false)
 	const [isResent, setIsResent] = useState([])
@@ -40,7 +39,11 @@ const Invites = ({ advisor }) => {
 	const [subject, setSubject] = useState(defaultSubject(advisor))
 
 	useEffect(() => {
-		fetch(`https://blockria.com/api/invites?advisorId=${advisor.idToken.payload.sub}`)
+		getPendingInvites()
+	}, [])
+
+	async function getPendingInvites() {
+		await fetch(`https://blockria.com/api/invites?advisorId=${advisor.idToken.payload.sub}`)
 			.then(response => response.json())
 			.then(newPendingInvites => {
 				let newIsResent = []
@@ -48,14 +51,27 @@ const Invites = ({ advisor }) => {
 				setIsResent(newIsResent)
 				setPendingInvites(newPendingInvites)
 			})
-			.catch(console.error)
-	}, [])
+			.catch(alert)
+	}
 
-	async function handleDelete() {
+	async function handleDelete(e, { clientEmailAddress }) {
+		if (isLoading) return
+
+		const deleteParams = { advisorId: advisor.idToken.payload.sub, clientEmailAddress }
+
 		await fetch('https://blockria.com/api/invites', {
 			method: 'DELETE',
-			headers: { 'Content-Type': 'application/json' }
+			headers: { 'Content-Type': 'application/json' },
+			body: JSON.stringify(deleteParams)
 		})
+
+		await getPendingInvites()
+	}
+
+	function handleSendAnotherInvite() {
+		setEmail('')
+		setIsLoading(false)
+		setIsSent(false)
 	}
 
 	async function handleSubmit(e) {
@@ -67,7 +83,10 @@ const Invites = ({ advisor }) => {
 			advisorName: `${advisor.idToken.payload.given_name} ${advisor.idToken.payload.family_name}`,
 			email,
 			message: message
-				.replace('[[Authorization Link]]', `<a href='${coinbaseURL}'>Authorization Link</a>`)
+				.replace(
+					'[[Authorization Link]]',
+					`<a href='${coinbaseURL}&state=["${advisor.idToken.payload.sub}","${email}"]'>Authorization Link</a>`
+				)
 				.replace(/\n/g, '<br>'),
 			subject
 		}
@@ -75,12 +94,12 @@ const Invites = ({ advisor }) => {
 		await emailjs
 			.send('service_1zzl4ah', 'template_zpo8s53', templateParams, 'joD1A0qHJYnXOOY4z')
 			.then(console.log)
-			.catch(console.error)
+			.catch(alert)
 
 		// [2.0] Save Records
 		const inviteOptions = {
 			advisorId: advisor.idToken.payload.sub,
-			clientEmailAddress: encodeURIComponent(email),
+			clientEmailAddress: email,
 			clientEmailLastSent: Date.now(),
 			clientEmailMessage: encodeURIComponent(message),
 			clientEmailSubject: subject
@@ -93,6 +112,8 @@ const Invites = ({ advisor }) => {
 
 		setIsLoading(false)
 		setIsSent(true)
+
+		await getPendingInvites()
 	}
 
 	function renderPendingInvite(pendingInvite, index) {
@@ -103,7 +124,7 @@ const Invites = ({ advisor }) => {
 
 		return (
 			<tr key={`Client Email ${pendingInvite.clientEmailAddress}`}>
-				<td>{decodeURIComponent(pendingInvite.clientEmailAddress)}</td>
+				<td>{pendingInvite.clientEmailAddress}</td>
 				<td>{daysSince ? (daysSince === 1 ? '1 Day Ago' : `${daysSince} Day Ago`) : 'Today'}</td>
 				{isResent[index] ? (
 					<td className='Bold'>Sent</td>
@@ -112,7 +133,9 @@ const Invites = ({ advisor }) => {
 						Resend Email
 					</td>
 				)}
-				<td className='Red'>Delete</td>
+				<td className='Red' onClick={e => handleDelete(e, pendingInvite)}>
+					Delete
+				</td>
 			</tr>
 		)
 	}
@@ -129,7 +152,7 @@ const Invites = ({ advisor }) => {
 		// [1.0] Send Email
 		const templateParams = {
 			advisorName: `${advisor.idToken.payload.given_name} ${advisor.idToken.payload.family_name}`,
-			email: decodeURIComponent(clientEmailAddress),
+			email: clientEmailAddress,
 			message: decodeURIComponent(clientEmailMessage)
 				.replace('[[Authorization Link]]', `<a href='${coinbaseURL}'>Authorization Link</a>`)
 				.replace(/\n/g, '<br>'),
@@ -139,7 +162,7 @@ const Invites = ({ advisor }) => {
 		await emailjs
 			.send('service_1zzl4ah', 'template_zpo8s53', templateParams, 'joD1A0qHJYnXOOY4z')
 			.then(console.log)
-			.catch(console.error)
+			.catch(alert)
 
 		// [2.0] Save Records
 		const inviteOptions = {
@@ -165,7 +188,14 @@ const Invites = ({ advisor }) => {
 				<form onSubmit={handleSubmit}>
 					<div className='Title'>New Invite</div>
 					<div>Client's Email</div>
-					<input onChange={e => setEmail(e.target.value)} placeholder='' required value={email} />
+					<input
+						autoComplete='No'
+						autoFocus
+						onChange={e => setEmail(e.target.value)}
+						required
+						type='text'
+						value={email}
+					/>
 					<div>Email Subject</div>
 					<input name='subject' onChange={e => setSubject(e.target.value)} required value={subject} />
 					<div>Email Body</div>
@@ -175,16 +205,16 @@ const Invites = ({ advisor }) => {
 						disabled={isLoading || isSent ? true : false}
 						style={isLoading || isSent ? { cursor: 'default', textDecoration: 'none' } : {}}
 						type='submit'
-						value={isSent ? 'Sent' : isLoading ? 'Loading...' : 'Send Email Invite'}
+						value={isSent ? 'Sent!' : isLoading ? 'Loading...' : 'Send Email Invite'}
 					/>
 					{isSent && (
-						<button className='SendAnotherEmail' onClick={() => window.location.reload()}>
+						<button className='SendAnotherEmail' onClick={() => handleSendAnotherInvite()}>
 							Send Another Invite
 						</button>
 					)}
 				</form>
 			</div>
-			{/* <table>
+			<table>
 				<caption>
 					<div className='Flex'>
 						<div className='Title'>Pending Invites</div>
@@ -204,7 +234,7 @@ const Invites = ({ advisor }) => {
 						<td>{pendingInvites.length} Total</td>
 					</tr>
 				</tfoot>
-			</table> */}
+			</table>
 		</div>
 	)
 }
